fixed the overflow of BeginTime/EndTime in displayin user events from Timeline.

+++ projections/gui/Timeline/UserEventWindow.java
@@ -79,10 +79,10 @@ public class UserEventWindow extends JFrame
 				switch (col) {
 				case 0:  
 					return ((UserEventObject)events_[index_][row]).getName();
-				case 1:  return new FormattedNumber((int)((UserEventObject)events_[index_][row]).BeginTime, format_);
-				case 2:  return new FormattedNumber((int)((UserEventObject)events_[index_][row]).EndTime, format_);
-				case 3:  return new FormattedNumber((int)((UserEventObject)events_[index_][row]).EndTime-
-						(int)((UserEventObject)events_[index_][row]).BeginTime, format_);
+				case 1:  return new FormattedNumber((long)((UserEventObject)events_[index_][row]).BeginTime, format_);
+				case 2:  return new FormattedNumber((long)((UserEventObject)events_[index_][row]).EndTime, format_);
+				case 3:  return new FormattedNumber((long)((UserEventObject)events_[index_][row]).EndTime-
+						(long)((UserEventObject)events_[index_][row]).BeginTime, format_);
 				default: return "ERROR";
 				}
 			}
